+++ controller/controller_test.go
@@ -11,6 +11,7 @@ import (
 )
 
 func TestRun(t *testing.T) {
+	t.Parallel()
 	resource1 := &testResource{
 		name: "resource1",
 	}